% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmarkov.R
\name{cemtool}
\alias{cemtool}
\title{Basic function to start the cemtool package}
\usage{
cemtool()
}
\description{
There are multiple software systems that can be used to build Markov models for cost-effectiveness analyses like TreeAge, Excel or R. Although there are numerous advantages to use R over the others, the biggest downside is the steep learning curve from R. The cemtool package aims to close this gap by introducing a step-by-step tool to guide users in building a default Markov models. The tool guides the user though the steps of the development of a Markov model and will present the final result using graphs and tables. The only prerequisite is that the user knows the structure of the model and the transition probabilities, no calculations or coding is required.
}
\author{

}
